+++ spec/rubocop/cop/style/space_around_keyword_spec.rb
@@ -131,6 +131,8 @@ describe RuboCop::Cop::Style::SpaceAroundKeyword do
   it_behaves_like 'missing after', 'super', 'super""', 'super ""'
   it_behaves_like 'accept after', '(', 'super(1)'
   it_behaves_like 'missing after', 'super', 'super{}', 'super {}'
+  it_behaves_like 'accept after', '(', 'defined?(1)'
+  it_behaves_like 'missing after', 'defined?', 'defined?1', 'defined? 1'
 
   it_behaves_like 'missing before', 'then', 'if ""then a end',
                   'if "" then a end'