+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [#2798](https://github.com/bbatsov/rubocop/pull/2798): `Rails/FindEach` cop works with `where.not`. ([@pocke][])
 * `Style/MultilineBlockLayout` can correct offenses which involve argument destructuring. ([@alexdowad][])
 * `Style/SpaceAroundKeyword` checks `super` nodes with no args. ([@alexdowad][])
+* `Style/SpaceAroundKeyword` checks `defined?` nodes. ([@alexdowad][])
 
 ### Bug fixes